feat: add route to delete books of collection and reactoring.

+++ src/domain/application/Collection/DeleteCollection/DeleteCollection.ts
@@ -4,6 +4,7 @@ import OutputBoundary from "@/domain/application/OutputBoundary";
 import UseCase from "@/domain/application/UseCase";
 import Repository from "@/domain/core/Repository";
 import DeleteCollectionOutputBoundary from "./DeleteCollectionOutputBoundary";
+import EntityNotFoundError from "../../Errors/EntityNotFoundError";
 
 export default class DeleteCollection
   implements UseCase<{ id: string }, boolean>
@@ -15,10 +16,13 @@ export default class DeleteCollection
   ): Promise<OutputBoundary<boolean>[]> {
     const { id } = inputData.get();
 
-    this.repository.delete({ id });
+    const deletedCollection = await this.repository.delete({ _id: id });
+    if (!deletedCollection) {
+      throw new EntityNotFoundError("Collection");
+    }
 
     const dbCollection: DBOutputCollectionData | null =
-      await this.repository.getOne({ id });
+      await this.repository.getOne({ _id: id });
 
     return [new DeleteCollectionOutputBoundary(dbCollection)];
   }
